Refactor some telemetry tests so they use same suite (#28635)
* Make dashboars test multicluster friendly and refactor some tests so they use same suite

fix

fix

enable loadbalancer for kind

enable loadbalancer for kind only in tracing tests

deploy gateway

make ingress multiple

debug

* fix stats test

* fix stats test

* remove verification of cluster-3
