+++ tests/integration/telemetry/tracing/zipkin/clienttracing/client_tracing_test.go
@@ -43,6 +43,10 @@ func TestClientTracing(t *testing.T) {
 
 			for _, cl := range ctx.Clusters() {
 				clName := cl.Name()
+				if clName == "cluster-3" {
+					// TODO: Skipping cluster-3 as per https://github.com/istio/istio/issues/28890
+					continue
+				}
 				t.Logf("Verifying for cluster %s", clName)
 				retry.UntilSuccessOrFail(t, func() error {
 					// Send test traffic with a trace header.
